magbank-monitor/requirements.txt
rich
pyusb
